added a way to customize messages and renamed message -> ChatMessage inside the folder itself
